[STEP1] Networking 분리, 추상화 	- ImageService, ImageServiceable DIP, SRP 적용
