+++ WeatherForecast/WeatherForecast/Controllers/WeatherListViewController.swift
@@ -35,7 +35,7 @@ class WeatherListViewController: UIViewController {
     }
     
     private func layoutWeatherListView() {
-        weatherListView = .init(weatherInfo: weatherInfo)
+        weatherListView = .init(weatherInfo: weatherInfo, imageService: ImageService())
         view.addSubview(weatherListView)
         weatherListView.translatesAutoresizingMaskIntoConstraints = false
         weatherListView.delegate = self
